Added Velocity Code to Swerve Drive
I also restructured the control loop for non-velocity control to be in the periodic loop as to keep the code structure consistent.

This commit also implements closed loop control of the swerve.

I removed updateOdometry as it should now be used anywhere else in the code thus the function call is not needed.